fix(android): fix an issue with android crashing due to a "no view found for id" exception
I'm not actually sure what the root cause is here. However, it seems to be due to some sort of
asynchronicity between the lifecycle stage of the hosting view and the fragment that replaces it.
The exception itself only occurs in certain situations and there's a bit of the "observer effect" at
play, as debugging a scenario which reproduces it will cause the issue to stop occurring. There
isn't much documentation to this effect currently the RN docs recommend replacing the view with your
fragment after "receiveCommand" has been called from the JS side. I implemented that and I could no
longer reproduce the issue on my device, however I'm not confident that it disappeared completely.