feat(android, ios, ontableofcontents): update onTableOfContents to return Locator[] for consistency
Links and Locators are not the same and need to be handled differently. To avoid complexity, the
library has been update to only deal with Locators and map all Links it returns to Locators.

BREAKING CHANGE: onTableOfContents now returns a `Locator[]` instead of a `Link[]`. Update your
`onTableOfContents` callbacks from `(toc: Link[]) => {...}` to `(toc: Locator[]) => {...}`.

https://github.com/5-stones/react-native-readium/issues/11